refactor: Merge all config builder into one ors config builder class
